doc: update doc build flow
* Add documentation build job to github actions
  * TODO: Remove from travis ci
* Update travis apt requirements
* Add cmake messages to help diagnosing when building documentation
  fails
* clean up build-documentation.sh script
  * remove dead code
  * terminate if installing python requirements fails
* clean up push-documentation.sh script
  * set user to "OPAE Bot", opae_robot@intel

+++ scripts/build-documentation.sh
@@ -1,21 +1,18 @@
 #!/bin/bash
 set -x
-COMMIT_ID="$(git rev-parse HEAD)"
-TEMP=$((git describe --exact-match $COMMIT_ID) 2>&1)
 PYTHON_VERSION=$(python3 --version | cut -d ' ' -f2)
-
-#if [[ $TEMP == *"fatal"* ]]; then
-#    echo "No new tag detected, hence builiding latest docs from master"
-#    git reset --hard origin/master
-#fi
 
 #install requirements
 python3 -m pip install --user -r doc/sphinx/requirements.txt
+if [ $? != 0 ];
+then
+	echo "failed to install Python requirements"
+	exit 1
+fi
 if [ -d mybuild_docs ];
 then
 	rm -rf mybuild_docs
 fi
-
 mkdir mybuild_docs
 pushd mybuild_docs
 export PYTHONPATH=$PWD/lib/python$PYTHON_VERSION
@@ -34,13 +31,3 @@ else
     echo "Spinx documentation build FAILED"
     exit 1
 fi
-
-
-
-# if [[ $TEMP != *"fatal"* ]]; then
-#     echo "New tag detected; Uploading documentation under new tag on opae.github.io"
-#     /bin/bash ../scripts/push-documentation.sh new_tag
-# else
-#     echo "No new tag detected; Latest documentation will be uploaded to opae.github.io"
-#     /bin/bash ../scripts/push-documentation.sh latest
-# fi
